# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    TaskHumanSummaryResponse:
      type: object
      required: [taskId, descriptionHash, language, record, stale, cached]
      properties:
        taskId:
          type: string
        descriptionHash:
          type: string
        language:
          type: string
          enum: [zh-CN, en]
        record:
          oneOf:
            - $ref: "#/components/schemas/TaskHumanSummaryRecord"
            - type: "null"
        stale:
          type: boolean
          description: true when record.descriptionHash differs from the current descriptionHash
        cached:
          type: boolean

    # ── Enums ──
    AgentRole:
      type: string
      enum: [ROUTA, CRAFTER, GATE, DEVELOPER]
...
    TaskStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, REVIEW_REQUIRED, COMPLETED, NEEDS_FIX, BLOCKED, CANCELLED]
...
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          const: true
...
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: assignedTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, objective]
              properties:
                title:
                  type: string
                objective:
                  type: string
                workspaceId:
                  type: string
                  default: "default"
                scope:
                  type: string
                acceptanceCriteria:
                  type: array
                  items:
                    type: string
                verificationCommands:
                  type: array
                  items:
                    type: string
                dependencies:
                  type: array
                  items:
                    type: string
                parallelGroup:
                  type: string
                boardId:
                  type: string
                columnId:
                  type: string
                position:
                  type: integer
                priority:
                  $ref: "#/components/schemas/TaskPriority"
                labels:
                  type: array
                  items:
                    type: string
                assignee:
                  type: string
                assignedProvider:
                  type: string
                assignedRole:
                  $ref: "#/components/schemas/AgentRole"
                assignedSpecialistId:
                  type: string
                assignedSpecialistName:
                  type: string
                createGitHubIssue:
                  type: boolean
                repoPath:
                  type: string
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteAllTasks
      summary: Delete all tasks for a workspace
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
  /api/tasks/{id}:
    get:
      operationId: getTask
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                objective:
                  type: string
                status:
                  $ref: "#/components/schemas/TaskStatus"
                boardId:
                  type: string
                columnId:
                  type: string
                position:
                  type: integer
                priority:
                  $ref: "#/components/schemas/TaskPriority"
                labels:
                  type: array
                  items:
                    type: string
                assignee:
                  type: string
                assignedProvider:
                  type: string
                assignedRole:
                  $ref: "#/components/schemas/AgentRole"
                assignedSpecialistId:
                  type: string
                assignedSpecialistName:
                  type: string
                repoPath:
                  type: string
                syncToGitHub:
                  type: boolean
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
  /api/tasks/{id}/status:
    post:
      operationId: updateTaskStatus
      summary: Update task status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
  /api/tasks/{id}/human-summary:
    get:
      operationId: getTaskHumanSummary
      summary: Read the cached human-readable card summary (never calls the model)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
            enum: [zh-CN, en]
            default: zh-CN
      responses:
        "200":
          description: Cached summary (record is null when none exists) plus the current description hash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskHumanSummaryResponse"
        "404":
          description: Task not found
    post:
      operationId: generateTaskHumanSummary
      summary: Return the cached summary when its hash matches the description, otherwise generate and cache
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  enum: [zh-CN, en]
                  default: zh-CN
                force:
                  type: boolean
                  description: Regenerate even when the cached hash matches
      responses:
        "200":
          description: Summary record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskHumanSummaryResponse"
        "404":
          description: Task not found
        "500":
          description: Model call failed or provider not configured
...
  /api/tasks/{id}/runs:
    get:
      operationId: listTaskRuns
      summary: List normalized execution runs for a task
      parameters:
        - name: id
          in: path
          required: true
          description: The task ID
          schema:
            type: string
      responses:
        "200":
          description: Task run ledger
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        "404":
          description: Task not found
...
  /api/tasks/{taskId}/changes:
    get:
      operationId: getTaskChanges
      summary: Get repository or worktree changes associated with a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task repository change summary
  /api/tasks/{taskId}/changes/file:
    get:
      operationId: getTaskChangeFile
      summary: Get diff for a single changed file associated with a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: true
          schema:
            type: string
        - name: previousPath
          in: query
          schema:
            type: string
      responses:
        "200":
          description: File diff payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  diff:
                    type: object
        "400":
          description: Missing path or repository unavailable
        "404":
          description: Task not found
  /api/tasks/{taskId}/changes/commit:
    get:
      operationId: getTaskChangeCommit
      summary: Get diff for a single commit associated with a task repository
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: sha
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Commit diff payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  diff:
                    type: string
        "400":
          description: Missing sha or repository unavailable
        "404":
          description: Task not found
...